feat: add swagger docs
