+++ src/main.ts
@@ -2,31 +2,22 @@ import * as process from 'process';
 
 import { NestFactory } from '@nestjs/core';
 import { ValidationPipe } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 
 import { AppModule } from './app.module';
+import { createSwaggerDoc } from './shared/lib';
 
 const PORT = process.env.PORT || 4000;
+
+const NODE_ENV = process.env.NODE_ENV || 'production';
 
 const bootstrap = async () => {
   const app = await NestFactory.create(AppModule);
 
   app.useGlobalPipes(new ValidationPipe());
 
-  const config = new DocumentBuilder()
-    .setTitle('Home Library Service')
-    .setDescription('Home music library service')
-    .setVersion('1.0')
-    .addTag('user')
-    .addTag('track')
-    .addTag('album')
-    .addTag('artist')
-    .addTag('favs')
-    .build();
-
-  const document = SwaggerModule.createDocument(app, config);
-
-  SwaggerModule.setup('api', app, document);
+  if (NODE_ENV === 'development') {
+    createSwaggerDoc(app);
+  }
 
   await app.listen(PORT, () => {
     console.log(`Server started on http://localhost:${PORT}`);
